+++ com/turquaz/cheque/ui/CheUIChequeOutPayrollCurrentUpdate.java
@@ -256,7 +256,7 @@ public class CheUIChequeOutPayrollCurrentUpdate extends org.eclipse.swt.widgets.
 					compChequeRoll.getTxtRollNo().getText().trim(),
 					compChequeRoll.getDatePicker1().getDate(),
 					chequeList,
-					EngBLCommon.CHEQUE_TRANS_OUT_CURRENT,
+					EngBLCommon.CHEQUE_TRANS_OUT_CURRENT.intValue(),
 					compChequeRoll.getBtnSumTotals().getSelection(),
 					EngBLCommon.getBaseCurrencyExchangeRate());
 	           EngUICommon.showMessageBox(getParent(),Messages.getString("CheUIChequeInPayroll.13"),SWT.ICON_INFORMATION); //$NON-NLS-1$
